feat: switch to phpstan level 3 (#373)

+++ admin/controllers/settings/views/backups/model.php
@@ -5,7 +5,7 @@ defined('CMSPATH') or die; // prevent unauthorized access
 
 function human_filesize($bytes, $decimals = 2) {
 	$sz = 'BKMGTP';
-	$factor = floor((strlen($bytes) - 1) / 3);
+	$factor = (int) floor((strlen($bytes) - 1) / 3);
 	return sprintf("%.{$decimals}f", $bytes / pow(1024, $factor)) . @$sz[$factor];
   }
